move FakeClock back to where it is
