+++ core/src/test/java/io/grpc/internal/TransportSetTest.java
@@ -58,7 +58,6 @@ import io.grpc.MethodDescriptor;
 import io.grpc.Status;
 import io.grpc.StringMarshaller;
 import io.grpc.internal.TestUtils.MockClientTransportInfo;
-import io.grpc.testing.FakeClock;
 
 import org.junit.After;
 import org.junit.Before;
